Add EconomyUtil class with money sending methods.
Implements a simple chainable type API for money transfer methods.
It uses the economy provider for the primitive withdraw/deposit methods and supports both banks and players.
The intention is to provide plugin developers a simple method to use for transactions between accounts, as many get it wrong.

Usage examples:
     EconomyUtil.send(23.2).fromPlayer("notch").toPlayer("jeb_")
     EconomyUtil.send(700).fromPlayer("ebenwert").toBank("Gringotts")

The following plugin can be used to test this feature: https://github.com/jastice/vault-tester