Make breaking change somewhat less breaking; add transfer methods to interface
Changing the Economy class to an abstract class broke all the plugins requesting the Economy interface.
This commit reverts it back to an interface, and adds the AbstractEconomy class instead.
This allows adding utility methods to the interface while maintaining backward compatibility.

Additionally, add 3 transfer methods to the interface for the most common money transfer use cases.
These are implemented in the abstract class

    EconomyResponse transfer(double amount, String playerFrom, String playerTo)
    EconomyResponse transferPlayerToBank(double amount, String playerFrom, String bankTo)
    EconomyResponse transferBankToPlayer(double amount, String bankFrom, String playerTo) {