Mark transitive dependencies as optional
Traditionally, Maven dependencies are marked as optional if they are not required to compile on top of the artifact, in this case, Vault. Usually though, since these dependencies are not marked as optional, the truly Mavenized dependencies (such as CraftConomy 3) will be sometimes included in the project structure, and unless excluded by the dependent plugin depending on Vault, those same dependencies will be leaked down the tree of dependencies.